backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.3
openpyxl==3.1.5
mysql-connector-python==9.0.0
python-multipart==0.0.9
pydantic==2.8.2
